bugfix: remove throwing an error if new remoteURL is different from the prevoious one (#268)
As presented in the #268 OP, this check did not work correctly, because it did not normalize URIs sent in a HTTP header into full URLs before checking.

Moreover, as explained in https://github.com/Palindrom/Palindrom/issues/268#issuecomment-527063823, there should NOT be such check on the client side. A new remote URL should not be considered an error. If the server has a problem in regenerating URLs, that problem should be fixed on the server, not detected as an error on the client. Replacing remote URLs can be completely valid, for example when:

1. the server decides to regenerate the session ID for security reasons
2. the server decides to redirect the client to another endpoint for performance reasons

It is up to the server to give an error when the client is trying to patch a wrong session due to some stale reference, session fixation, etc.